obtain repos does not need a queue
